requirements.txt
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.2
scikit-learn==1.4.2
category_encoders==2.6.4
optuna==3.6.1
catboost==1.2.7
kaleido==1.2.0
plotly==6.5.0

u8darts==0.30.0
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
pytorch-lightning==2.2.5
torchmetrics==1.3.2

mlflow==2.12.1
